hw/scripts/interrupt_map/requirements.txt
jinja2==3.1.4
pydantic==2.7.3
pyyaml==6.0.1
